Note API CRUD project 1 complete

+++ controllers/noteController.js
@@ -28,12 +28,12 @@ const updateNotes = async (req, res) => {
         const { title, content, isPinned } = req.body
 
         // ------------ Find and update 
-        if (noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request") 
+        if (!noteId || noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
         const note = await noteSchema.findOneAndUpdate({ _id: noteId }, { title, content, isPinned }, { new: true })
         if (!note) return res.status(404).send("Note couldn't found")
 
         // ------------- Success 
-        res.status(200).send("Note updated successfully", note)
+        res.status(200).send("Note updated successfully")
     } catch (error) {
         console.log(error)
     }
@@ -44,12 +44,12 @@ const deletNotes = async (req, res) => {
     try {
         const { noteId } = req.params
 
-        if (noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
+        if (!noteId || noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
         const note = await noteSchema.findOneAndDelete({ _id: noteId })
         if (!note) return res.status(400).send("Note couldn't Found")
 
         // ------------- Success 
-        res.status(204).send("Note Deleted successfully")
+        res.status(200).send("Note Deleted successfully")
     } catch (error) {
         console.log(error)
     }
@@ -59,8 +59,6 @@ const deletNotes = async (req, res) => {
 // ======================= Get All notes 
 const getAllNotes = async (req, res) => {
     try {
-        if (noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
-
         const note = await noteSchema.find({})
         if (!note) return res.status(400).send("Note couldn't Found")
 
@@ -76,7 +74,7 @@ const getAllNotes = async (req, res) => {
 const getSingleNote = async (req, res) => {
     try {
         const { noteId } = req.params
-        if (noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
+        if (!noteId || noteId.length > 24 || noteId.length < 24) return res.status(400).send("Invalid request")
         const note = await noteSchema.findOne({ _id: noteId })
         if (!note) return res.status(400).send("Note couldn't Found")
